+++ lib/dart_mc_ping.dart
@@ -1,8 +1,8 @@
 library dart_mc_ping;
 
 import 'package:dart_mc_ping/dart_mc_client.dart';
+import 'package:dart_mc_ping/logger.dart';
 import 'package:dart_mc_ping/model/status_response.dart';
-import 'package:logging/logging.dart';
 
 /// Ping a Minecraft server and return the server status.
 Future<StatusResponse> ping(String host, {int port = 25565}) async {
@@ -10,7 +10,7 @@ Future<StatusResponse> ping(String host, {int port = 25565}) async {
   await client.connect();
 
   final statusResponse = await client.ping();
-  Logger.root.info('$host:$port -> ${statusResponse.ms} ms');
+  logger.info('$host:$port -> ${statusResponse.ms} ms');
 
   await client.close();
   return statusResponse;