+++ lib/logger.dart
@@ -1,9 +1,10 @@
 import 'package:logging/logging.dart';
 
-/// Configure the log levels. Useful for debugging purposes.
+final Logger logger = Logger.detached("MC_PING");
+
 void configureLogger({Level level = Level.OFF}) {
-  Logger.root.level = level;
-  Logger.root.onRecord.listen((record) {
+  logger.level = level;
+  logger.onRecord.listen((record) {
     final level = '${record.level.name.padRight(5)}';
     print('[$level] ${record.time} | ${record.message}');
   });